requirements.txt
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
dnspython==1.16.0
eventlet==0.25.0
Flask==1.0.2
Flask-SocketIO==4.2.0
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-engineio==3.9.0
python-socketio==4.3.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.5.0
Werkzeug==0.15.5
